feat(ui): display average rating with stars on main page and detail

+++ backend/guideproject/src/main/java/academy/mediasoft/team/guideproject/dto/RatingDto.java
@@ -11,7 +11,7 @@ public record RatingDto(
         @Min(1)
         @Max(5)
         @NotNull(message = "Оценка должна быть поставлена")
-        Integer grade,
+        Double grade,
         LocalDateTime createdAt,
         Long landmarkId
 ) { }
